LayoutTests/imported/w3c/web-platform-tests/css/css-fonts/animations/font-variation-settings-composition-expected.txt

CONSOLE MESSAGE: line 14: ReferenceError: Can't find variable: test_composition

Harness Error (FAIL), message = ReferenceError: Can't find variable: test_composition
